+++ src/main.ts
@@ -1,6 +1,7 @@
 import { ActivityType } from "discord.js";
 import { registeredHandlers } from "src/decorators/handler";
 import { Server } from "src/entities/server";
+import { ServerMember } from "./entities/server-member";
 import { dataSource } from "src/init/data-source";
 import { client, generateInvite, withClient } from "src/init/discord";
 import { cyclePresence } from "src/presence/cycle-presence";
@@ -8,8 +9,8 @@ import { getHierarchy, getHierarchyRole } from "src/roles/hierachy/hierachy";
 import { roleByName } from "src/roles/role-by-name";
 import { cyclicIterator } from "src/util";
 import "./commands";
-import { ServerMember } from "./entities/server-member";
 import "./services";
+import { addMemberEntry, setMemberLeft } from "./members/member-entries";
 export async function initialise() {
   await dataSource.initialize();
   const client = await withClient();
@@ -75,9 +76,7 @@ client.on("guildMemberRemove", async member => {
 
   // await removeCustomPronounRole(member as any);
   if(member.user.bot) return;
-  const userEntry = await dataSource.getRepository(ServerMember).findOne({ where: { discordId: member.user.id } });
-  userEntry.leftAt = new Date();
-  await dataSource.getRepository(ServerMember).save(userEntry);
+  await setMemberLeft(member);
 });
 
 client.on("guildMemberAdd", async member => {
@@ -94,20 +93,7 @@ client.on("guildMemberAdd", async member => {
     if (!role) await member.roles.add(hierarchy[0]);
     console.log(member.user.username, "ist dem server beigetreten");
     // TODO: welcome messages
-    const guild  = await dataSource
-        .getRepository(Server)
-        .findOne({ where: { discordId: member.guild.id } });
-
-    const existingUser = await dataSource.getRepository(ServerMember).findOne({ where: { discordId: member.user.id } });
-    if (existingUser){
-      existingUser.leftAt = null;
-      await dataSource.getRepository(ServerMember).save(existingUser);
-      return;
-    }
-    const userEntry : Partial<ServerMember> = { discordId: member.user.id, username: member.user.username, discriminator: member.user.discriminator, avatarUrl: member.user.avatarURL(), guild };
-    userEntry.discordId = member.user.id;
-    await dataSource.getRepository(ServerMember).save(userEntry);
-
+    await addMemberEntry(member);
   }
 });
