+++ src/members/member-entries.ts
@@ -1,0 +1,26 @@
+import { GuildMember, PartialGuildMember } from "discord.js";
+import { Server } from "../entities/server";
+import { ServerMember } from "../entities/server-member";
+import { dataSource } from "../init/data-source";
+
+export const addMemberEntry = async (member: GuildMember) => {
+    const guild  = await dataSource
+        .getRepository(Server)
+        .findOne({ where: { discordId: member.guild.id } });
+
+    const existingUser = await dataSource.getRepository(ServerMember).findOne({ where: { discordId: member.user.id } });
+    if (existingUser){
+      existingUser.leftAt = null;
+      await dataSource.getRepository(ServerMember).save(existingUser);
+      return;
+    }
+    const userEntry : Partial<ServerMember> = { discordId: member.user.id, username: member.user.username, discriminator: member.user.discriminator, avatarUrl: member.user.avatarURL(), guild };
+    userEntry.discordId = member.user.id;
+    await dataSource.getRepository(ServerMember).save(userEntry);
+}
+
+export const setMemberLeft = async (member: GuildMember | PartialGuildMember) => {
+  const userEntry = await dataSource.getRepository(ServerMember).findOne({ where: { discordId: member.user.id } });
+  userEntry.leftAt = new Date();
+  await dataSource.getRepository(ServerMember).save(userEntry);
+}
